backend/requirements.txt
fastapi
uvicorn
python-multipart
openai
python-dotenv
sqlalchemy
gTTS
SpeechRecognition
pydub
requests

# Optional TTS providers (install as needed)
# google-cloud-texttospeech  # For Google Cloud TTS
# azure-cognitiveservices-speech  # For Azure TTS
